refactor: replace the Cartlist.vue with MultiStepForm.vue
